Config: Rename config library to make it work with Desktop
