+++ client/config/index.js
@@ -1,8 +1,7 @@
-// DO NOT USE ES6 MODULES  OR IT WILL BREAK wp-desktop COMPATIBILITY
 /**
  * Internal dependencies
  */
-var createConfig = require( 'lib/config' );
+import createConfig from 'lib/create-config';
 
 /**
  * Manages config flags for various deployment builds
@@ -12,5 +11,5 @@ if ( 'undefined' === typeof window || ! window.configData ) {
 	throw new ReferenceError( 'No configuration was found: please see client/config/README.md for more information' );
 }
 
-module.exports = createConfig( window.configData );
+export default createConfig( window.configData );
