+++ client/lib/create-config/index.js
