+++ client/sections.js
@@ -7,7 +7,8 @@ const path = require( 'path' );
 /**
  * Internal dependencies
  */
-const sections = require( 'wordpress-com' );
+const config = require( 'config' );
+const sections = require( config( 'project' ) );
 const extensions = require( 'extensions' );
 
 const extensionSections = extensions.map( extension => {
