+++ server/config/index.js
@@ -3,7 +3,7 @@
  */
 const configPath = require( 'path' ).resolve( __dirname, '..', '..', 'config' );
 const parser = require( './parser' );
-const createConfig = require( 'lib/config' );
+const createConfig = require( 'lib/create-config' );
 
 const { serverData: data, clientData } = parser( configPath, {
 	env: process.env.CALYPSO_ENV || process.env.NODE_ENV || 'development',
